\encoding{UTF-8}


\name{CreateInputsPerturb}
\alias{CreateInputsPerturb}


\title{Generation of ensemble model inputs for data assimilation}

\usage{
CreateInputsPerturb(FUN_MOD, DatesR, Precip = NULL, PotEvap = NULL, NbMbr, Seed = NULL)

}

\arguments{

\item{FUN_MOD}{[function] hydrological model function (i.e. \code{\link{RunModel_GR5J}})}

\item{DatesR}{[POSIXct] vector of dates}

\item{Precip}{(optional) [numeric] time series of total precipitation [mm/d]}

\item{PotEvap}{(optional) [numeric] time series of potential evapotranspiration [mm/d]}

\item{NbMbr}{[numeric] number of ensemble members}

\item{Seed}{(optional) [numeric] seed of random number generator}
}

\value{

\emph{InputsPert}{[list] object of class \emph{InputsModel} containing the ensembles of perturbed model inputs required to perform data assimilation:}

  \tabular{ll}{                                                                                         
     \emph{$DatesR } \tab [POSIXlt] vector of dates
                          \cr\cr
     \emph{$Precip } \tab [matrix, dim(Nt, NbMbr)] ensemblist time series of perturbed total precipitation [mm/d] (with Nt the length of the period; generated only if the precipitation time series is provided as a function argument).  
                          \cr\cr
     \emph{$PotEvap} \tab [matrix, dim(Nt, NbMbr)] ensemblist time series of perturbed potential evapotranspiration [mm/d] (generated only if the time series of potential evapotranspiration is provided as a function argument).                        
         }                                                                                      
}

\description{
Function which perturbs the model inputs to generate probabilistic meteorological forcings required to perform ensemble-based data assimilation.
}

\details{
The function generates an ensemble for each meteorological variable provided as an argument. 
\cr
Probabilistic forcing/s is/are generated by stochastically perturbing the meteorological variable/s through a multiplicative stochastic noise, according to the methodology proposed by Clark et al. (2008).   
\cr
The random perturbations are provided through a first-order autoregressive model relying on a fractional error parameter equal to 0.65 and a temporal decorrelation length of 1 day for rainfall and 2 days for potential evapotranspiration.
\cr
In order to ensure reproducible results, \emph{Seed} can be set to fix the randomness in the generation of perturbations. 
\cr
For further details, see the references section.
}

\examples{
library(airGRdatassim)

## loading catchment data
data(L0123001, package = "airGR")

## preparation of the InputsModel object
InputsModel <- CreateInputsModel(FUN_MOD = RunModel_GR5J, DatesR = BasinObs$DatesR, 
                                 Precip = BasinObs$P, PotEvap = BasinObs$E)
                                 
## run period selection
IndRun <- seq(which(format(BasinObs$DatesR, format = "\%Y-\%m-\%d")=="2006-01-01"),
              which(format(BasinObs$DatesR, format = "\%Y-\%m-\%d")=="2006-01-31"))

## preparation of perturbed meteorological ensemble            
InputsPert <- CreateInputsPerturb(FUN_MOD = RunModel_GR5J, DatesR = BasinObs$DatesR[IndRun],
                                  Precip = BasinObs$P[IndRun],
                                  NbMbr = 100)               

## results preview
rangePrecip <- apply(InputsPert$Precip, MARGIN = 1, FUN = range) 
plot(x = InputsPert$DatesR, y = rowMeans(InputsPert$Precip),
     ylim = range(rangePrecip),
     type = "l", col = "red", lwd = 2,
     main = "Precipitation ensemble",
     xlab = "Time [day]", ylab = "Total precipitation [mm/day]",
     panel.first = polygon(x = c(as.numeric(InputsPert$DatesR), rev(as.numeric(InputsPert$DatesR))),
                           y = c(rangePrecip[1, ], rev(rangePrecip[2, ])),
                           col = "pale turquoise", border = NA))
}

\author{
Gaia Piazzi, Olivier Delaigue
}

\references{
Clark, M. P., Rupp, D. E., Woods, R. A., Zheng, X., Ibbitt, R. P., Slater, A. G. et al. (2008). Hydrological data assimilation with the ensemble Kalman filter: Use of streamflow observations to update states in a distributed hydrological model. Advances in Water Resources, 31(10), 1309-1324.
\cr
Piazzi, G., Thirel, G.,  Perrin, C. and Delaigue, O. (2020). Sequential data assimilation for streamflow forecasting: assessing the sensitivity to uncertainties and updated variables of a conceptual hydrological model. Water Resources Research. (submitted).
}

\seealso{
\code{\link{RunModel_DA}}
}
